Add separate `platform` role (#6)
Fix the `docker_platform` role so that it works correctly when
more than one platform is specified in molecule.yml.

Move non-provider-specific platform configuration to a `platform` role. This is a breaking change for this version, as functions of `docker_platform` are now provided by platform.

Replace any references to `docker_platform` in playbooks using this collection with `platform`, options stay the same.